fixs public IP import bug wrt idle timeout & some minor refactoring
